chore(rebase): adapt to master's avro package move + convertValueToEngineType contract
Rebased the stack onto current apache/hudi master, which since the branch's base:
- moved HoodieAvroUtils and AvroRecordContext from org.apache.hudi.avro to
  org.apache.hudi.common.avro (update the two imports), and
- redefined convertValueToEngineType to return the engine-native STORAGE form
  (e.g. an unwrapped Spark UTF8String), splitting the comparison-wrapping into
  ensureComparability/convertOrderingValueToEngineType. So the synthesized union
  changed-columns value now uses convertValueToEngineType (was the interim
  convertPartitionValueToEngineType, which no longer exists on master).

Verified against rebased master: hudi-common Oracle unit tests (62) green;
TestOracleDebeziumV9ReadMerge IT 6/6 green; checkstyle/rat/scalastyle clean.
