+++ hudi-common/src/main/java/org/apache/hudi/common/model/debezium/OracleDebeziumAvroPayload.java
@@ -18,7 +18,7 @@
 
 package org.apache.hudi.common.model.debezium;
 
-import org.apache.hudi.avro.HoodieAvroUtils;
+import org.apache.hudi.common.avro.HoodieAvroUtils;
 import org.apache.hudi.common.model.HoodieRecord;
 import org.apache.hudi.common.model.OverwriteWithLatestAvroPayload;
 import org.apache.hudi.common.model.PartialUpdateAvroPayload;
